# docs/swagger.yaml
basePath: /api/v1
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  models.Author:
    properties:
      about:
        example: it's info about myself
        type: string
      avatar:
        example: filename.jpeg
        type: string
      email:
        example: admin@mail.ru
        type: string
      is_author:
        example: true
        type: boolean
      username:
        example: admin
        type: string
    required:
    - email
    - is_author
    - username
    type: object
  models.EmptyStruct:
    type: object
  models.Post:
    properties:
      id:
        example: 1
        type: integer
      img:
        example: path/to/image.jpeg
        type: string
      text:
        example: some text
        type: string
      title:
        example: some title
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - img
    - text
    - title
    type: object
  models.ResponseImage:
    properties:
      img_path:
        example: /path/to/image.jpeg
        type: string
    required:
    - img_path
    type: object
  models.Subscription:
    properties:
      author_id:
        example: 1
        type: integer
      author_subscription_id:
        example: 1
        type: integer
      subscriber_id:
        example: 2
        type: integer
    required:
    - author_id
    - author_subscription_id
    - subscriber_id
    type: object
...
host: vdonate.ml:8080
info:
  contact:
    email: zeronethunter2001@gmail.com
    name: VDonate Support
  termsOfService: http://swagger.io/terms/
  title: VDonate API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get posts with provided filters
      operationId: get_posts
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Posts were successfully received
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Get posts
      tags:
      - posts
    post:
      consumes:
      - multipart/form-data
      description: Create post by user cookie
      operationId: create_post
      parameters:
      - example: some text
        in: formData
        name: text
        required: true
        type: string
      - example: some title
        in: formData
        name: title
        required: true
        type: string
      - description: Uploaded file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Post was successfully created
          schema:
            $ref: '#/definitions/models.ResponseImage'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Not a creator of post
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error / failed to create post
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Create post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete post by post id
      operationId: delete_post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post was successfully deleted
          schema:
            $ref: '#/definitions/models.EmptyStruct'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Not a creator of post
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get single post by post id
      operationId: get_post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post was successfully received
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Get single post
      tags:
      - posts
    put:
      consumes:
      - multipart/form-data
      description: Update post by post id
      operationId: update_post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - example: some text
        in: formData
        name: text
        required: true
        type: string
      - example: some title
        in: formData
        name: title
        required: true
        type: string
      - description: Uploaded file
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Post was successfully updated
          schema:
            $ref: '#/definitions/models.ResponseImage'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: No session provided
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Not a creator of post
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal error / failed to create image
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Update post
      tags:
      - posts
  /subscribers:
    delete:
      consumes:
      - application/json
      description: Unsubscribe user from author
      operationId: delete_subscriber
      parameters:
      - description: Subscription info with required UserID, AuthorID and Subscription
          ID
        in: body
        name: Subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: Subscriber was successfully unsubscribed
          schema:
            $ref: '#/definitions/models.EmptyStruct'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Not deleted
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Unsubscribe
      tags:
      - subscribers
    post:
      consumes:
      - application/json
      description: Subscribe user to author with paid subscription
      operationId: create_subscriber
      parameters:
      - description: Subscription info with required UserID, AuthorID and Subscription
          ID
        in: body
        name: Subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully subscribed
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Not created
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Subscribe
      tags:
      - subscribers
  /subscribers/{author_id}:
    get:
      description: Request to server for subscriptions of requested `Author`
      operationId: get_subscribers
      parameters:
      - description: Author ID
        in: path
        name: author_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get list of author subscribers
          schema:
            items:
              $ref: '#/definitions/models.Author'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Returns subscribers by author ID
      tags:
      - subscribers
...
securityDefinitions:
  ApiKeyAuth:
    description: Authorization via Cookie
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"